all user-specific validations completed in userform

+++ src/Components/UserDataForm.js
@@ -1,45 +1,57 @@
 import { useUser } from "@clerk/clerk-react";
 import { Box, Button, Container, TextField, Typography } from "@mui/material";
-import React, { useEffect, useRef, useState } from "react";
+import React, { useEffect, useState } from "react";
 import { v4 as uuidv4 } from 'uuid';
+import { setName, setEmail, setPhone, setAddress, setUserID } from "../redux/slices/UserDataSlice";
+import { useDispatch, useSelector } from "react-redux";
+import { alreadyAUser } from "../utils/utils";
 
 export default function UserDataForm() {
 
     const { user } = useUser();
     const userId = user.id; 
+
+    const dispatch = useDispatch();
+    const { name, email, phone, address } = useSelector(state => state.userdata);
     
     const [formUpdated, setFormUpdated] = useState(false);
-    const [isPreviousUser, setIsPreviousUser] = useState(false);
-
-    const nameRef = useRef(null);
-    const emailRef = useRef(null);
-    const phoneRef = useRef(null);
-    const addressRef = useRef(null);
 
     const handleSubmit = (e) => {
+
         e.preventDefault();
-        const name = nameRef.current.querySelector('input').value;
-        const email = emailRef.current.querySelector('input').value;
-        const phone = phoneRef.current.querySelector('input').value;
-        const address = addressRef.current.querySelector('textarea').value;
         
         if (name === '' || email === '' || phone === '' || address === '') {
             alert('Kindly fill all the details');
             return;
         }
 
-        const userDataObj = { name: name, email: email, phone: phone, address: address, userID: userId ? userId : uuidv4() };
-        
+        const userid = userId ? userId : uuidv4();
+        dispatch(setUserID(userid))
+
+        const userDataObj = { name: name, email: email, phone: phone, address: address, userID: userid };
+
         const usersData = JSON.parse(localStorage.getItem('usersData'));
         if (usersData === null) {
             const usersData = [];
             usersData.push(userDataObj);
             localStorage.setItem('usersData', JSON.stringify(usersData));
         } else {
-            usersData.push(userDataObj);
-            localStorage.setItem('usersData', JSON.stringify(usersData));
+            if (alreadyAUser(usersData,userId)) {
+                console.log('yes');
+                const updatedUsersData = usersData.map(obj => {
+                    if (obj.userID === userId) {
+                        obj = {...userDataObj}
+                    }
+                    return obj;
+                })
+                localStorage.setItem('usersData', JSON.stringify(updatedUsersData));
+            } else {
+                usersData.push(userDataObj);
+                localStorage.setItem('usersData', JSON.stringify(usersData));
+            }
         }
         alert('Details updated successfully!');
+
     }
 
     useEffect(() => {
@@ -62,8 +74,14 @@ export default function UserDataForm() {
     useEffect(() => {
         const usersData = JSON.parse(localStorage.getItem('usersData'));
         if (usersData !== null) {
-            const oldUser = usersData.some(obj => obj.userID === userId);
-            if (oldUser) setIsPreviousUser(true);
+            if (alreadyAUser(usersData, userId)) {
+                const currentUser = usersData.filter(obj => obj.userID === userId);
+                const { name, email, phone, address } = currentUser[0];
+                dispatch(setName(name));
+                dispatch(setEmail(email));
+                dispatch(setPhone(phone));
+                dispatch(setAddress(address));
+            }
         }
     }, [])
    
@@ -74,8 +92,7 @@ export default function UserDataForm() {
                     <Typography component="h1" variant="h5">FORM</Typography>
                     <Box component="form" sx={{ mt: 3 }}>
                     <TextField 
-                        value={isPreviousUser && isPreviousUser}
-                        ref={nameRef}
+                        value={name}
                         margin="normal"
                         required
                         fullWidth
@@ -84,11 +101,14 @@ export default function UserDataForm() {
                         name="name"
                         autoComplete="name"
                         autoFocus
-                        onChange={() => setFormUpdated(true)}                    
+                        onChange={(e) => {
+                            setFormUpdated(true)
+                            dispatch(setName(e.target.value))
+                        }}
                     />
                     
                     <TextField
-                        ref={emailRef}
+                        value={email}
                         margin="normal"
                         required
                         fullWidth
@@ -97,10 +117,13 @@ export default function UserDataForm() {
                         name="email"
                         autoComplete="email"
                         type="email"
-                        onChange={() => setFormUpdated(true)}
+                        onChange={(e) => {
+                            setFormUpdated(true)
+                            dispatch(setEmail(e.target.value))
+                        }}
                     />
                     <TextField
-                        ref={phoneRef}
+                        value={phone}
                         margin="normal"
                         required
                         fullWidth
@@ -109,10 +132,13 @@ export default function UserDataForm() {
                         name="phone"
                         autoComplete="phone"
                         type="number"
-                        onChange={() => setFormUpdated(true)}
+                        onChange={(e) => {
+                            setFormUpdated(true)
+                            dispatch(setPhone(e.target.value))
+                        }}
                     />
                     <TextField
-                        ref={addressRef}
+                        value={address}
                         margin="normal"
                         required
                         fullWidth
@@ -122,7 +148,10 @@ export default function UserDataForm() {
                         autoComplete="address"
                         multiline
                         rows={4}
-                        onChange={() => setFormUpdated(true)}
+                        onChange={(e) => {
+                            setFormUpdated(true)
+                            dispatch(setAddress(e.target.value))
+                        }}
                     />
                     <Button
                         onClick={handleSubmit}
